Update index.html
Change games subsite link from games.yiru.me to games.rmwinslow.com

The dual website thing was worth trying out, but it was just too much of a headache. Maybe I'll switch ownership of the repo or something in the future (which would require setting up redirects and so on) but for now, I'll just update the links to ignore the mirror repo on the bacontime account.